refactor: move sendMiniGameStatistics
