+++ src/app/redux/features/audiocall/audiocallSlice.ts
@@ -4,12 +4,8 @@ import { IGameQuestion } from '../../../interfaces/IGameQuestion';
 import { IGameQuestionArray } from '../../../interfaces/IGameQuestionArray';
 import { IPaginatedArray } from '../../../interfaces/IPaginatedArray';
 import { IResultData } from '../../../interfaces/IResultData';
-import { IUserWord } from '../../../interfaces/IUserWord';
 import { IWord } from '../../../interfaces/IWord';
-import { TokenProvider } from '../../../services/TokenProvider';
-import { UserWordService } from '../../../services/UserWordService';
 import { WordService } from '../../../services/WordService';
-import { executePromisesSequentially } from '../../../utils/executePromisesSequentially';
 
 interface AudiocallPageState {
   questions: IGameQuestionArray | undefined;
@@ -75,35 +71,6 @@ const getQuestions = async (args: LevelAndPageArgs) => WordService.getWordsByGro
 export const getQuestionsAction = createAsyncThunk<IPaginatedArray<IWord>, LevelAndPageArgs>(
   'audiocall/getQuestions',
   getQuestions,
-);
-
-const sendMiniGameStatistics = async (args: GameResults) => {
-  const isExpired = TokenProvider.checkIsExpired();
-  const userId = TokenProvider.getUserId();
-
-  if (isExpired || !userId) {
-    return false;
-  }
-
-  // eslint-disable-next-line arrow-body-style
-  const promiseFunction = (item: IResultData) => {
-    return () => UserWordService.setWordStatistic(userId, item.questionData.id, item.isCorrect);
-  };
-
-  const setWordStatisticPromises = args.results.map(promiseFunction);
-
-  await executePromisesSequentially<IUserWord>(setWordStatisticPromises)
-    .then()
-    .catch((error) => {
-      console.error('Error:', error);
-    });
-
-  return true;
-};
-
-export const sendMiniGameStatisticsAction = createAsyncThunk<boolean, GameResults>(
-  'audiocall/sendMiniGameStatistics',
-  sendMiniGameStatistics,
 );
 
 export const audiocallSlice = createSlice({
